Added Tabs to Tasks

+++ app/Views/tab/tab.php
@@ -30,11 +30,13 @@
                         echo "<tr>";
                             echo('<td>' . $tab['Name'] . '</td>');
                             echo('<td>' . $tab['Beschreibung'] . '</td>');
-                            echo(' <td class="text-end">
-                                <a href="tab_edit?tab='.$tab['ID'].'"><i class="fa-regular fa-pen-to-square"></i></a>
-                                <a href="tab_delete?tab='.$tab['ID'].'"><i class="fa-regular fa-trash-can"></i></a>
-                            </td>
-                        </tr>');
+                            if(isset($_SESSION['ID'])) {
+                                echo(' <td class="text-end">
+                                <a href="tab_edit?tab=' . $tab['ID'] . '"><i class="fa-regular fa-pen-to-square"></i></a>
+                                <a href="tab_delete?tab=' . $tab['ID'] . '"><i class="fa-regular fa-trash-can"></i></a>
+                                </td>');
+                            }
+                            echo('</tr>');
                         }
                         ?>
                     </tbody>
